# app/server/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CommentRequest:
    properties:
      bookId:
        type: integer
      chapterId:
        type: integer
      content:
        type: string
      parentId:
        type: integer
      replyToId:
        type: integer
    required:
    - bookId
    - chapterId
    - content
    type: object
  dto.CommentResponse:
    properties:
      bookId:
        type: integer
      chapterId:
        type: integer
      content:
        type: string
      createBy:
        type: integer
      createTime:
        type: string
      deptId:
        type: integer
      id:
        type: integer
      likeCount:
        type: integer
      liked:
        type: boolean
      ownerId:
        type: integer
      parentId:
        type: integer
      readerId:
        type: integer
      readerName:
        type: string
      replies:
        items:
          $ref: '#/definitions/dto.CommentResponse'
        type: array
      replyToId:
        type: integer
      replyToName:
        type: string
      status:
        type: string
      updateBy:
        type: integer
      updateTime:
        type: string
    type: object
  dto.CommentSearch:
    properties:
      chapterId:
        type: integer
      current:
        type: integer
      keyword:
        type: string
      parentId:
        type: integer
      size:
        type: integer
    type: object
...
  dto.LikeRequest:
    properties:
      targetId:
        type: integer
      targetType:
        enum:
        - "1"
        - "2"
        - "3"
        type: string
    required:
    - targetId
    - targetType
    type: object
  dto.LikeResponse:
    properties:
      count:
        type: integer
      liked:
        type: boolean
    type: object
  dto.LikeStatusRequest:
    properties:
      targets:
        items:
          $ref: '#/definitions/dto.LikeTarget'
        minItems: 1
        type: array
    required:
    - targets
    type: object
  dto.LikeTarget:
    properties:
      targetId:
        type: integer
      targetType:
        type: string
    type: object
...
  dto.NoteRequest:
    properties:
      bookId:
        type: integer
      chapterId:
        type: integer
      content:
        type: string
      endOffset:
        type: integer
      highlightColor:
        type: string
      noteType:
        enum:
        - "1"
        - "2"
        - "3"
        type: string
      selectedText:
        type: string
      startOffset:
        type: integer
      visibility:
        enum:
        - "1"
        - "2"
        type: string
    required:
    - bookId
    - noteType
    type: object
  dto.NoteResponse:
    properties:
      bookId:
        type: integer
      bookTitle:
        type: string
      chapterId:
        type: integer
      chapterTitle:
        type: string
      content:
        type: string
      createBy:
        type: integer
      createTime:
        type: string
      endOffset:
        type: integer
      highlightColor:
        type: string
      id:
        type: integer
      likeCount:
        type: integer
      liked:
        type: boolean
      noteType:
        type: string
      readerId:
        type: integer
      readerName:
        type: string
      selectedText:
        type: string
      startOffset:
        type: integer
      updateBy:
        type: integer
      updateTime:
        type: string
      visibility:
        type: string
    type: object
  dto.NoteSearch:
    properties:
      bookId:
        type: integer
      chapterId:
        type: integer
      current:
        type: integer
      keyword:
        type: string
      noteType:
        type: string
      size:
        type: integer
    type: object
...
  dto.PageResponse:
    properties:
      current:
        type: integer
      records: {}
      size:
        type: integer
      total:
        type: integer
    type: object
...
  dto.ReviewRequest:
    properties:
      bookId:
        type: integer
      content:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      title:
        type: string
    required:
    - bookId
    - content
    type: object
  dto.ReviewResponse:
    properties:
      bookId:
        type: integer
      bookTitle:
        type: string
      content:
        type: string
      createBy:
        type: integer
      createTime:
        type: string
      deptId:
        type: integer
      id:
        type: integer
      likeCount:
        type: integer
      liked:
        type: boolean
      ownerId:
        type: integer
      rating:
        type: integer
      readerId:
        type: integer
      readerName:
        type: string
      status:
        type: string
      title:
        type: string
      updateBy:
        type: integer
      updateTime:
        type: string
    type: object
  dto.ReviewSearch:
    properties:
      bookId:
        type: integer
      current:
        type: integer
      keyword:
        type: string
      size:
        type: integer
      status:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 小说阅读平台后端 API
  title: Boread API
  version: "1.0"
paths:
...
  /api/book/comment:
    post:
      consumes:
      - application/json
      parameters:
      - description: 评论参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CommentResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建章节评论
      tags:
      - book-comment
  /api/book/comment/{id}:
    delete:
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除章节评论
      tags:
      - book-comment
    get:
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CommentResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取评论详情
      tags:
      - book-comment
  /api/book/comment/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 搜索参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CommentSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PageResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 分页查询评论
      tags:
      - book-comment
...
  /api/book/like/count/{targetType}/{targetId}:
    get:
      parameters:
      - description: 目标类型
        in: path
        name: targetType
        required: true
        type: string
      - description: 目标ID
        in: path
        name: targetId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  format: int64
                  type: integer
              type: object
      security:
      - BearerAuth: []
      summary: 查询点赞数
      tags:
      - like
  /api/book/like/status:
    post:
      consumes:
      - application/json
      parameters:
      - description: 查询参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LikeStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: boolean
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 批量查询点赞状态
      tags:
      - like
  /api/book/like/toggle:
    post:
      consumes:
      - application/json
      parameters:
      - description: 点赞参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LikeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LikeResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 切换点赞
      tags:
      - like
  /api/book/note:
    post:
      consumes:
      - application/json
      parameters:
      - description: 笔记参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.NoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.NoteResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建笔记/划线
      tags:
      - reader-note
  /api/book/note/{id}:
    delete:
      parameters:
      - description: 笔记ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除笔记/划线
      tags:
      - reader-note
    get:
      parameters:
      - description: 笔记ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.NoteResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取笔记详情
      tags:
      - reader-note
    put:
      consumes:
      - application/json
      parameters:
      - description: 笔记ID
        in: path
        name: id
        required: true
        type: integer
      - description: 笔记参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.NoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.NoteResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新笔记/划线
      tags:
      - reader-note
  /api/book/note/book/{bookId}:
    get:
      parameters:
      - description: 书籍ID
        in: path
        name: bookId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.NoteResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取某本书的公开笔记
      tags:
      - reader-note
  /api/book/note/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 搜索参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.NoteSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PageResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 分页查询笔记
      tags:
      - reader-note
...
  /api/book/review:
    post:
      consumes:
      - application/json
      parameters:
      - description: 书评参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ReviewResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建书评
      tags:
      - book-review
  /api/book/review/{id}:
    delete:
      parameters:
      - description: 书评ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除书评
      tags:
      - book-review
    get:
      parameters:
      - description: 书评ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ReviewResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取书评详情
      tags:
      - book-review
    put:
      consumes:
      - application/json
      parameters:
      - description: 书评ID
        in: path
        name: id
        required: true
        type: integer
      - description: 书评参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ReviewResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新书评
      tags:
      - book-review
  /api/book/review/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 搜索参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ReviewSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PageResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 分页查询书评
      tags:
      - book-review
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"